test: Initial testing with jest.
