+++ package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "build": "tsc",
     "lint": "eslint",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "keywords": [
     "pouchdb"
@@ -17,12 +17,15 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@eslint/js": "9.20.0",
+    "@types/jest": "29.5.14",
     "@types/pouchdb": "6.4.2",
     "@types/pouchdb-find": "7.3.3",
     "eslint": "9.20.1",
     "eslint-config-prettier": "10.0.1",
     "eslint-plugin-prettier": "5.2.3",
+    "jest": "29.7.0",
     "prettier": "3.5.1",
+    "ts-jest": "29.2.5",
     "typescript": "5.7.3",
     "typescript-eslint": "8.24.0"
   },
