+++ src/heartdb.test.ts
@@ -1,0 +1,14 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview Jasmine tests for HeartDB.
+ */
+
+import { HeartDB } from "../src/heartdb";
+
+describe("HeartDB", () => {
+  it("should be a constructor function", () => {
+    expect(HeartDB).toBeInstanceOf(Function);
+  });
+});
